moved the event rendering logic into a separate component
